[#13] Update URL when drilling down the treemap
We're just doing the first part of the integration. Right now, if you click on
a treemap's tile, the URL will change, and you can go back to the previous
state. But if you access the URL with some state, it won't reflect on the
treemap. Yet.

I've also removed stuff related to entities, that were useful in INESC, but not
in this project. Not now, at least.