[#13] Treemap navigation and we just load the current drilldown
I've had to change a few things. First of all, we just have the IDs on the URL.
This makes the URL impossible to read, but eases my life (for now).

When the user enters in a page, for example:
  /#/treemap/28/845/73000/73108/40

We will have to query OpenSpending to get each drilldown's label, so we can
create the breadcrumb. When we do this, it'll be easy to change these URLs to
have slugs as well. But until then, let's use IDs.

As per just loading the current drilldown, I had to do this so we could
recreate a state from the URL. I had, basically, two options.

One was to always load the entire treemap first, and then parse the URL to
find which state the user is in, and simulate the user's "clicks" to get there.
It would solve the above issue, of having to query OS to get each drilldown's
label (if we load everything, we would have them). But it sounded quite complex,
and I thought the treemap would flicker as we're loading everything and ONLY
THEN drilling down. Also, this would be impossible to do with the amount of
drilldowns we want to use, as there'll be too much data.

The other option was to always only load the current drilldown. Every time you
click on a tile, we would fetch the required data. So, when the user requests
to see the treemap of function's 28, subfunction 845, we would load the treemap
with these cuts already. It saves us from the need to hack the treemap to be
able to simulate the user's clicks to recreate the state. As a GREAT bonus, we
solve the "OMG! There's too much data" problem :)

There're a few known bugs that I added as FIXMEs.

(And probably a few unkown hiding around)