[#9] Add breadcrumb service to gather the labels in OpenSpending
To understand why we need this, consider this: a user hits the page:

  /#/treemap/2013/10-saude/301-atencao-basica

How do we, having just the URL, recreate the breadcrumb on the page? We loose
information when we slugify a string (casing and accents, for example), so we
can't recreate the original string from the slug. Our options are to live with
it, and try our best (i.e. turn atencao-basica into ATENCAO BASICA). Or we can
hit OpenSpending to get the correct labels.

I did the latter.

Looking back now, we could have changed our slug function do to turn something
like "Atenção Básica" into "aten'c~ao-b'asica". From there, we could recreate
the accents, just missing the up/lower casing. I could live with that, as we'll
probably display it all in uppercase anyway. This would probably be better, as
we save a hit in OpenSpending. But well, it's done now. On the positive side,
we've got cleaner slugs :P