[#17] Add contextual data download button
Instead of downloading just the data necessary to recreate the graphs shown in
the current screen, as we at first thought of doing, we use the current state
as a filter.

For example, if you're seeing the page for the year 2012 and funcao Saude,
instead of just getting the distribution of money for every subfuncao inside
Health in that year (i.e. what we show in the treemap), you would get
everything we've got, filtered by year 2012 and function Health. You would get
every subfuncao, orgao, uo, etc. This, obviously, includes what's used to build
the current screen, but you get much more data.

We decided that it was better, as the user can get everything related to the
state he's in (2012, Saude), and it's not too much data. For example,
everything for 2012 is less than 6 MB.