[#16] Fix problem where clicking on links added its ID to the breadcrumb
The problem was that we were simply following the link, without going through
the breadcrumb. So, when the routing service tried to update the breadcrumb
from the URL, it couldn't get the label (as it's not possible to get that just
from the URL). As we didn't had the label, we used the ID.

Now, instead of simply following the link, we actually call updateState() with
the element related to that link, full with its id, label and type. This way,
everything works as expected.

I had a problem adding angular directives to the dataTables. To solve them, I
had to $compile each table's row whenever they're change with the scope, so
Angular got a change to check for directives and do its magic.

Next time, it might be worthy checking http://angular-ui.github.io/ng-grid/.
This is like a dataTable wrote natively in Angular. It looks interesting, and
could simplify our code, but I couldn't find how to hook a callback on onClick
an element. And it felt overkill, as it's not a big change. With dataTables we
can reuse more code/scss, so I prefered to stay with it. For now.