[#16] Fix link slugs when first entering the page
I was (re)generating the links based on when scope.cuts changed. When you enter
the page, the code parses the URL and populates the breadcrumb and scope.cuts,
then tries to gather the breadcrumb's labels from OpenSpending async. Before
that happened, we already created the links and, as we didn't had the labels,
they became something like:

  #/treemap/10/301-atencao-basica

Now we watch on the breadcrumb. So, whenever we finish gathering the labels
from OpenSpending, it'll change and trigger a change in the URLs. Now we have
nice slugs :-)

  #/treemap/10-saude/301-atencao-basica