[#41] Updates breadcrumb labels when updating the state as well
There are two situations when we need to update the breadcrumb's labels. First,
whenever we open the page. As we only have the slugs from the URL, and we can't
recreate the labels from it, we need to hit OpenSpending to get them.

The other way is when you're navigating to a state (say "2013 > SAUDE"), then
go back to a previous state (i.e. "2013"). By this time, our code already
updated the breadcrumb, removing the FUNCAO and leaving only ANO (so we lost
the label). Finally, you press your browser's BACK button, to go to the
original state.

In this case, we can only get the breadcrumb from the URL (/2013/10-saude), and
we have the same problem of the first situation again: having just the slug,
and not the label. So we have to hit OpenSpending one more time. To avoid this call,
I built a cache.

In all other situations (i.e. drilling down by clicking on the treemap), we
already have the labels, because the treemap itself brings everything when it
loads.

The cache might be a bit overkill, but it was simple enough and we're already
doing too many calls to OpenSpending. Maybe it would be good to refactor that
out to an Angular Service in the future.