[#18] Add navigation by years on the percentual change bars
To do so, I had to change in basically every place where scope.year used to be
an integer, to make it a string. This was needed because the radio button's
values are strings and, as they're controlled by AngularJS, there's no easy way
(AFAIK) to convert them to integers before setting the scope in the value.

All in all, this makes our lives easier, as we don't have to keep casting it.
We only converted so we could turn "10-saude" to 10 in the first place, and
that doesn't matter for the year, so we're good.

Maybe it would be good to change in the routing service to do something like:

  parseInt(id).toString()

So it would work the same way for every drilldown in the breadcrumb, but let's
KISS for now.