[#18] Fix algorithm to calculate the percentage variation between years
We were calculating using the absolute value. For example, if 2011 the budget
was 1 billion, and in 2012 it was 2 billion, we say that it was increased by
100%. That's wrong, because we're not deflating.

What we really meant to do (and are doing now) is to compare values relative to
the total. Say that the budget was 1 billion in 2011, and the total budget in
that year was 1 trillion, so it was 0.1% in 2011. In 2012, say that the budget
was 1.5 billions, and the total budget was 2 trillion, so it was 0.075% in that
year. The number we want is this change from 0.1% in 2011 to 0.075% in 2012: a
25% decrease.

As we are always getting the budgets relative to the total budget, we don't
have to take the inflation in consideration.