[#20] Set the treemap tiles' colors to grey in the canvas
To make the choropleth treemap, We're not colorizing the tiles themselves, but
divs that float over it. This works very well, because we're able to use CSS to
define the colors, which keeps the code neat and separated, and allows reuse
whenever we create the legends.

But, when we're drilling down into the treemap, changing the state, the spinner
appears, and we can see the real colors of the treemap from underneath it. I'm
changing it to grey to make it more discrete.