[#21] Fix sorting by the budget execution percentage
The problem was that we tried to get the percentual value by doing
"executadoCol.innerHTML". It worked before, as the <td> only had this value.
But now that we've added a bunch of <span> to create the bars, it doesn't work
anymore.

So we have to use jQuery to find the .meter-horizontal-label element and get
its value. Or, in cases where the Autorizado value is 0 (i.e., the percentual
would be divided by zero, giving us Infinity%), we simpy put a blank string in
the td. But, for sorting purposes, we consider these lines as -Infinity, so
they'll always be the last ones when sorting by ascending order.