[#6] Include the templates into the HTML
This should make the overall experience of using the site faster, as you don't
have to do a request for every partial a page uses. I would prefer to have used
grunt-angular-template, which creates a .js file with all the templates,
instead of inlining it into the HTML. This would allow us to let the user cache
the templates, which would make things even faster. We can't allow she to cache
index.html, because it changes a bit. At least not for long.

I created an issue in grunt-angular-templates' repository to try to solve the
issue I had. If they (or I) do, we can change to use it.

https://github.com/ericclemmons/grunt-angular-templates/issues/50